+++ test/ecdsa.test.js
@@ -1,18 +1,21 @@
-require('terst')
+var BigInteger = require('bigi');
+var secureRandom = require('secure-random');
+
+var sha256 = require('sha256');
+var ecurve = require('ecurve')
+var ecparams = ecurve.getECParams('secp256k1')
 
 var ECDSA = require('../');
-var sha256 = require('sha256');
-var secureRandom = require('secure-random');
-var ecparams = require('ecurve-names')('secp256k1');
-var BigInteger = require('bigi');
+
+require('terst')
 
 describe('ECDSA()', function() {
   it('should create the ecdsa object with the proper curve', function() {
     var ecparams = require('ecurve-names')('secp256k1');
     var ecdsa = new ECDSA(ecparams);
-    EQ (ecdsa.ecparams.getG().toString(16), ecparams.getG().toString(16));
-    EQ (ecdsa.ecparams.getN().toString(16), ecparams.getN().toString(16));
-    EQ (ecdsa.ecparams.getH().toString(16), ecparams.getH().toString(16));
+    EQ (ecdsa.ecparams.g.toString(16), ecparams.getG().toString(16));
+    EQ (ecdsa.ecparams.n.toString(16), ecparams.getN().toString(16));
+    EQ (ecdsa.ecparams.h.toString(16), ecparams.getH().toString(16));
   })
 })
 
@@ -39,7 +42,7 @@ describe('- verify()', function() {
       var privKey = BigInteger.fromByteArrayUnsigned(randArr)
       var ecdsa = new ECDSA(ecparams);
       //var privKey = ecdsa.getBigRandom(ecparams.getN())
-      var pubPoint = ecparams.getG().multiply(privKey)
+      var pubPoint = ecparams.g.multiply(privKey)
       var pubKey = pubPoint.getEncoded(true) //true => compressed
       var msg = "hello world!"
       var shaMsg = sha256(msg)
@@ -58,7 +61,7 @@ describe('+ verify()', function() {
       
       ECDSA.ecparams = ecparams;
       //var privKey = ecdsa.getBigRandom(ecparams.getN())
-      var pubPoint = ecparams.getG().multiply(privKey)
+      var pubPoint = ecparams.g.multiply(privKey)
       var pubKey = pubPoint.getEncoded(false) //true => compressed
       var msg = "hello world!"
       var shaMsg = sha256(msg)
@@ -73,7 +76,7 @@ describe('+ verify()', function() {
       var randArr = secureRandom(32, {array: true})
       var privKey = BigInteger.fromByteArrayUnsigned(randArr)
       //var privKey = ecdsa.getBigRandom(ecparams.getN())
-      var pubPoint = ecparams.getG().multiply(privKey)
+      var pubPoint = ecparams.g.multiply(privKey)
       var pubKey = pubPoint.getEncoded(true) //true => compressed
       var msg = "hello world!"
       var shaMsg = sha256(msg)